requirements.txt
fastapi==0.104.0
uvicorn==0.23.2
pydantic==2.4.2
python-multipart==0.0.6
moviepy==1.0.3
Pillow==10.0.1
google-generativeai==0.3.0
python-dotenv==1.0.0
google-auth==2.23.3
ffmpeg-python==0.2.0
aiofiles==23.2.1